% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_namespace.R
\name{check_namespace}
\alias{check_namespace}
\title{Read a package NAMESPACE and check for dangerous imports}
\usage{
check_namespace(pkg_path, dangerous_imports = character(0))
}
\arguments{
\item{pkg_path}{path to package source tree}

\item{dangerous_imports}{character vector of dangerous items to find}
}
\description{
Given a path to a package source tree, return a list of Imports (both whole
packages and fully qualified references).
}
\examples{
\dontrun{
check_namespace("../testevil", c("processx", "sys", "processx::run"))
}
}
