% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dangerous_imports}
\alias{dangerous_imports}
\title{Dangerous Imports in Namespace}
\usage{
dangerous_imports(additional_dangerous_imports = character(0))
}
\arguments{
\item{additional_dangerous_imports}{character vector of user-defined
dangerous imports}
}
\value{
character vector of all imports considered dangerous
}
\description{
Character vector of all imports considered dangerous to be used with
\code{\link{check_namespace}}.
}
\examples{
dangerous_imports()
}
