% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_calls.R
\name{summarize_system_calls}
\alias{summarize_system_calls}
\title{summarize occurences of system calls in project}
\usage{
summarize_system_calls(path)
}
\arguments{
\item{path}{character, relative or absolute path for local code directory}
}
\value{
data.frame with a row for each system call
}
\description{
Summarize and show occurences of system calls in the R files in a given
folder. There could be more occurences. These calls should be further checked
for maliciousness.
}
\examples{
\dontrun{
# git clone git@github.com:ropenscilabs/testevil.git
summarize_system_calls("testevil")
}
}
