% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{system_calls}
\alias{system_calls}
\title{Functions making System Calls}
\usage{
system_calls(additional_system_calls = character(0))
}
\arguments{
\item{additional_system_calls}{character vector of user-defined
system calls}
}
\value{
character vector of all system calls
}
\description{
Character vector of functions making system calls to check for in R source
files.
}
\examples{
system_calls("exec_background")
}
