% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_namespace.R
\name{check_namespace}
\alias{check_namespace}
\title{Read a package NAMESPACE and check for dangerous imports}
\usage{
check_namespace(pkg_path, imports_to_flag = dangerous_imports())
}
\arguments{
\item{pkg_path}{path to package source tree}

\item{imports_to_flag}{character vector of dangerous items to find}
}
\description{
Given a path to a package source tree, return a data.frame of Imports (both whole
packages and fully qualified references).
}
\examples{
\dontrun{
check_namespace("../testevil")
check_namespace(
  "../testevil",
  dangerous_imports(additional_dangerous_imports = "sys::exec_background")
)
}
}
