% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defender-package.R
\docType{package}
\name{defender-package}
\alias{defender}
\alias{defender-package}
\title{defender: Check Package for Potential Security Violations}
\description{
Check an R package for potential security risks and violations via static code analysis.
}
\author{
\strong{Maintainer}: Ildi Czeller \email{czeildi@gmail.com}

}
\keyword{internal}
