% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defender-package.R
\docType{package}
\name{defender-package}
\alias{defender}
\alias{defender-package}
\title{defender: Check Package for Potential Security Violations}
\description{
Check an R package for potential security risks and violations via static code analysis.
}
\author{
\strong{Maintainer}: Ildiko Czeller \email{czeildi@gmail.com}

Authors:
\itemize{
  \item Karthik Ram \email{karthik.ram@gmail.com} ( 0000-0002-0233-1757)
  \item Bob Rudis \email{bob@rud.is} (0000-0001-5670-2640)
  \item Kara Woo
}

}
\keyword{internal}
